FUND-2391 DRC null check on inhoud field (#165)
null check on inhoud field

+++ src/OneGround.ZGW.Documenten.Web/Handlers/v1/5/CreateEnkelvoudigInformatieObjectCommandHandler.cs
@@ -231,6 +231,11 @@ public class CreateEnkelvoudigInformatieObjectCommandHandler
 
     private async Task RollbackDocumentJustAddedToDmsAsync(string urnDocument)
     {
+        if (string.IsNullOrEmpty(urnDocument))
+        {
+            return;
+        }
+
         try
         {
             await DocumentService.DeleteDocumentAsync(new DocumentUrn(urnDocument));
